update output function of inference script and add method to app and minor update to training model

+++ sagemaker/sentiment/deploy.py
@@ -19,6 +19,7 @@ import os
 import time
 import tarfile
 import shutil
+import csv
 
 import boto3
 import glob
@@ -52,17 +53,17 @@ INSTANCE_TYPE = "ml.t2.medium"
 # Model artifact paths (relative to this script)
 SCRIPT_DIR: str = os.path.dirname(p=os.path.abspath(path=__file__))
 MODEL_PATHS: list[str] = [
-    os.path.join(SCRIPT_DIR, "inference.py"),
-    os.path.join(SCRIPT_DIR, "pytorch_model.bin"),
-    os.path.join(SCRIPT_DIR, "tokenizer.json"),
-    os.path.join(SCRIPT_DIR, "requirements.txt"),
-    os.path.join(SCRIPT_DIR, "special_tokens_map.json"),
-    os.path.join(SCRIPT_DIR, "tokenizer_config.json"),
-    os.path.join(SCRIPT_DIR, "config.json"),
-    os.path.join(SCRIPT_DIR, "vocab.txt"),
-    os.path.join(SCRIPT_DIR, "sentiment_columns.json"),
-    os.path.join(SCRIPT_DIR, "sentiment_encoders.json"),
-    os.path.join(SCRIPT_DIR, "sentiment_schema.json")
+    os.path.join(SCRIPT_DIR, "model/inference.py"),
+    os.path.join(SCRIPT_DIR, "model/sentiment_columns.json"),
+    os.path.join(SCRIPT_DIR, "model/sentiment_encoders.json"),
+    os.path.join(SCRIPT_DIR, "model/sentiment_schema.json"),
+    os.path.join(SCRIPT_DIR, "model/requirements.txt"),
+    os.path.join(SCRIPT_DIR, "model/pytorch_model.bin"),
+    os.path.join(SCRIPT_DIR, "model/tokenizer.json"),
+    os.path.join(SCRIPT_DIR, "model/special_tokens_map.json"),
+    os.path.join(SCRIPT_DIR, "model/tokenizer_config.json"),
+    os.path.join(SCRIPT_DIR, "model/config.json"),
+    os.path.join(SCRIPT_DIR, "model/vocab.txt"),
 ]
 
 def shutdown_threads() -> None:
@@ -163,6 +164,10 @@ def validate_model_artifacts() -> None:
 def package_model(tar_path: str) -> str:
     """Package model artifacts into a tar.gz for SageMaker."""
     validate_model_artifacts()
+    
+    if not os.path.exists(path="exported_model"):
+        os.makedirs(name=f"{SCRIPT_DIR}/exported_model", exist_ok=True)
+        
     print(f"Packaging model to {tar_path}...")
     with tarfile.open(name=tar_path, mode="w:gz") as tr:
         for filepath in MODEL_PATHS:
@@ -198,16 +203,14 @@ def download_s3_folder() -> None:
     s3 = boto3.client("s3", region_name=REGION)
 
     paginator = s3.get_paginator("list_objects_v2")
-
     
+    os.makedirs(name=f"{SCRIPT_DIR}/exported_model", exist_ok=True)
     
     for page in paginator.paginate(Bucket=S3_BUCKET, Prefix=MODEL_PREFIX):
         for obj in page.get("Contents", []):
             s3_key = obj["Key"]
             rel_path = s3_key[len(MODEL_PREFIX):].lstrip("/")
             local_path = os.path.join(os.path.dirname(__file__), rel_path)
-            
-            os.makedirs(name=f"{SCRIPT_DIR}/exported_model", exist_ok=True)
             
             s3.download_file(S3_BUCKET, s3_key, local_path)
 
@@ -445,6 +448,11 @@ def test_endpoint() -> None:
         "customer_service_count": 0,
         "customer_issue_history": 0,
     }
+    
+    with open(file=os.path.join(os.path.dirname(__file__), "call_transcripts.csv")) as f:
+        csv_file = csv.DictReader(f)
+
+        test_payload = next(csv_file)
 
     print("Sending test prediction...")
     response = runtime.invoke_endpoint(
@@ -466,20 +474,22 @@ if __name__ == "__main__":
     parser.add_argument("--delete", action="store_true", help="Delete the endpoint")
     parser.add_argument("--test", action="store_true", help="Test the live endpoint")
     args = parser.parse_args()
-
+    
     if args.delete:
         delete_endpoint()
     elif args.test:
         test_endpoint()
     else:
-        download_s3_folder()
         valid_checks()
         tar_path = os.path.join(SCRIPT_DIR, "model.tar.gz")
         package_model(tar_path=tar_path)
         s3_uri = upload_to_s3(tar_path=tar_path)
+        download_s3_folder()
         create_endpoint(s3_uri=s3_uri)
-        os.remove(path="./model.tar.gz")
-        os.remove(path="./mlflow.db")
+        if os.path.isdir(s="./model.tar.gz"):
+            os.remove(path="./model.tar.gz")
         if os.path.isdir(s="./model"):
             shutil.rmtree(path="./model")
+        if os.path.isdir(s="./exported_model"):
+            os.rmdir(path="./exported_model")
     atexit.register(shutdown_threads)
